requirements.txt
asgiref==3.5.2
autopep8==1.6.0      
Django==4.0.6        
django-tinymce==3.4.0
Pillow==9.2.0        
psycopg2==2.9.3      
pycodestyle==2.8.0   
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1